# Serializer performance improvements on property name lookup