--- a/src/System.Text.Json/src/System/Text/Json/Serialization/JsonClassInfo.cs
+++ b/src/System.Text.Json/src/System/Text/Json/Serialization/JsonClassInfo.cs
@@ -5,6 +5,7 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.Diagnostics;
+using System.Linq;
 using System.Reflection;
 using System.Runtime.InteropServices;
 using System.Text.Json.Serialization;
@@ -18,9 +19,13 @@ namespace System.Text.Json
         // The length of the property name embedded in the key (in bytes).
         private const int PropertyNameKeyLength = 6;
 
-        private readonly List<PropertyRef> _propertyRefs = new List<PropertyRef>();
+        // The limit to how many property names from the JSON are cached before using PropertyCache.
+        private const int PropertyNameCountCacheThreshold = 64;
 
-        // Cache of properties by first ordering. Use an array for highest performance.
+        // The properties on a POCO keyed on property name.
+        public volatile Dictionary<string, JsonPropertyInfo> PropertyCache;
+
+        // Cache of properties by first JSON ordering. Use an array for highest performance.
         private volatile PropertyRef[] _propertyRefsSorted = null;
 
         internal delegate object ConstructorDelegate();
@@ -33,47 +38,37 @@ namespace System.Text.Json
         // If enumerable, the JsonClassInfo for the element type.
         internal JsonClassInfo ElementClassInfo { get; private set; }
 
+        public JsonSerializerOptions Options { get; private set; }
+
         public Type Type { get; private set; }
 
         internal void UpdateSortedPropertyCache(ref ReadStackFrame frame)
         {
-            // Todo: on classes with many properties (about 50) we need to switch to a hashtable for performance.
-            // Todo: when using PropertyNameCaseInsensitive we also need to use the hashtable with case-insensitive
-            // comparison to handle Turkish etc. cultures properly.
-
-            Debug.Assert(_propertyRefs != null);
-
-            // Set the sorted property cache. Overwrite any existing cache which can occur in multi-threaded cases.
-            if (frame.PropertyRefCache != null)
+            // Check if we are trying to build the sorted cache.
+            if (frame.PropertyRefCache == null)
             {
-                HashSet<PropertyRef> cache = frame.PropertyRefCache;
-
-                // Add any missing properties. This creates a consistent cache count which is important for
-                // the loop in GetProperty() when there are multiple threads in a race conditions each generating
-                // a cache for a given POCO type but with different property counts in the json.
-                while (cache.Count < _propertyRefs.Count)
-                {
-                    for (int iProperty = 0; iProperty < _propertyRefs.Count; iProperty++)
-                    {
-                        PropertyRef propertyRef = _propertyRefs[iProperty];
-                        // Cache the missing property or override the existing property.
-                        cache.Add(propertyRef);
-                    }
-                }
-
-                Debug.Assert(cache.Count == _propertyRefs.Count);
-                if (_propertyRefsSorted == null || _propertyRefsSorted.Length < cache.Count)
-                {
-                    _propertyRefsSorted = new PropertyRef[cache.Count];
-                }
-                cache.CopyTo(_propertyRefsSorted);
-                frame.PropertyRefCache = null;
+                return;
             }
+
+            List<PropertyRef> newList;
+            if (_propertyRefsSorted != null)
+            {
+                newList = new List<PropertyRef>(_propertyRefsSorted);
+                newList.AddRange(frame.PropertyRefCache);
+                _propertyRefsSorted = newList.ToArray();
+            }
+            else
+            {
+                _propertyRefsSorted = frame.PropertyRefCache.ToArray();
+            }
+
+            frame.PropertyRefCache = null;
         }
 
         internal JsonClassInfo(Type type, JsonSerializerOptions options)
         {
             Type = type;
+            Options = options;
             ClassType = GetClassType(type, options);
 
             CreateObject = options.MemberAccessorStrategy.CreateConstructor(type);
@@ -83,9 +78,11 @@ namespace System.Text.Json
             {
                 case ClassType.Object:
                     {
-                        var propertyNames = new HashSet<string>(StringComparer.Ordinal);
+                        PropertyInfo[] properties = type.GetProperties(BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic);
 
-                        foreach (PropertyInfo propertyInfo in type.GetProperties(BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic))
+                        Dictionary<string, JsonPropertyInfo> cache = CreatePropertyCache(properties.Length);
+
+                        foreach (PropertyInfo propertyInfo in properties)
                         {
                             // Ignore indexers
                             if (propertyInfo.GetIndexParameters().Length > 0)
@@ -99,17 +96,18 @@ namespace System.Text.Json
                             {
                                 JsonPropertyInfo jsonPropertyInfo = AddProperty(propertyInfo.PropertyType, propertyInfo, type, options);
 
-                                Debug.Assert(jsonPropertyInfo.NameUsedToCompareAsString != null);
-
                                 // If the JsonPropertyNameAttribute or naming policy results in collisions, throw an exception.
-                                if (!propertyNames.Add(jsonPropertyInfo.NameUsedToCompareAsString))
+                                if (cache.ContainsKey(jsonPropertyInfo.NameAsString))
                                 {
                                     ThrowHelper.ThrowInvalidOperationException_SerializerPropertyNameConflict(this, jsonPropertyInfo);
                                 }
 
-                                jsonPropertyInfo.ClearUnusedValuesAfterAdd();
+                                cache.Add(jsonPropertyInfo.NameAsString, jsonPropertyInfo);
                             }
                         }
+
+                        // Set as a unit to avoid concurrency issues.
+                        PropertyCache = cache;
 
                         DetermineExtensionDataProperty();
                     }
@@ -118,10 +116,10 @@ namespace System.Text.Json
                 case ClassType.Dictionary:
                     {
                         // Add a single property that maps to the class type so we can have policies applied.
-                        JsonPropertyInfo policyProperty = AddPolicyProperty(type, options);
+                         AddPolicyProperty(type, options);
 
                         // Use the type from the property policy to get any late-bound concrete types (from an interface like IDictionary).
-                        CreateObject = options.MemberAccessorStrategy.CreateConstructor(policyProperty.RuntimePropertyType);
+                        CreateObject = options.MemberAccessorStrategy.CreateConstructor(PolicyProperty.RuntimePropertyType);
 
                         // Create a ClassInfo that maps to the element type which is used for (de)serialization and policies.
                         Type elementType = GetElementType(type, parentType: null, memberInfo: null, options: options);
@@ -143,9 +141,13 @@ namespace System.Text.Json
                     }
                     break;
                 case ClassType.Value:
+                    // Add a single property that maps to the class type so we can have policies applied.
+                    AddPolicyProperty(type, options);
+                    break;
                 case ClassType.Unknown:
                     // Add a single property that maps to the class type so we can have policies applied.
                     AddPolicyProperty(type, options);
+                    PropertyCache = new Dictionary<string, JsonPropertyInfo>();
                     break;
                 default:
                     Debug.Fail($"Unexpected class type: {ClassType}");
@@ -171,14 +173,12 @@ namespace System.Text.Json
 
         private JsonPropertyInfo GetPropertyThatHasAttribute(Type attributeType)
         {
-            Debug.Assert(_propertyRefs != null);
+            Debug.Assert(PropertyCache != null);
 
             JsonPropertyInfo property = null;
 
-            for (int iProperty = 0; iProperty < _propertyRefs.Count; iProperty++)
+            foreach (JsonPropertyInfo jsonPropertyInfo in PropertyCache.Values)
             {
-                PropertyRef propertyRef = _propertyRefs[iProperty];
-                JsonPropertyInfo jsonPropertyInfo = propertyRef.Info;
                 Attribute attribute = jsonPropertyInfo.PropertyInfo.GetCustomAttribute(attributeType);
                 if (attribute != null)
                 {
@@ -194,37 +194,29 @@ namespace System.Text.Json
             return property;
         }
 
-        internal JsonPropertyInfo GetProperty(JsonSerializerOptions options, ReadOnlySpan<byte> propertyName, ref ReadStackFrame frame)
+        internal JsonPropertyInfo GetProperty(ReadOnlySpan<byte> propertyName, ref ReadStackFrame frame)
         {
-            // If we should compare with case-insensitive, normalize to an uppercase format since that is what is cached on the propertyInfo.
-            if (options.PropertyNameCaseInsensitive)
-            {
-                string utf16PropertyName = JsonHelpers.Utf8GetString(propertyName);
-                string upper = utf16PropertyName.ToUpperInvariant();
-                propertyName = Encoding.UTF8.GetBytes(upper);
-            }
-
-            ulong key = GetKey(propertyName);
             JsonPropertyInfo info = null;
 
-            // First try sorted lookup.
-            int propertyIndex = frame.PropertyIndex;
-
             // If we're not trying to build the cache locally, and there is an existing cache, then use it.
-            bool hasPropertyCache = frame.PropertyRefCache == null && _propertyRefsSorted != null;
-            if (hasPropertyCache)
+            if (_propertyRefsSorted != null)
             {
+                ulong key = GetKey(propertyName);
+
+                // First try sorted lookup.
+                int propertyIndex = frame.PropertyIndex;
+
                 // This .Length is consistent no matter what json data intialized _propertyRefsSorted.
                 int count = _propertyRefsSorted.Length;
                 if (count != 0)
                 {
-                    int iForward = propertyIndex;
-                    int iBackward = propertyIndex - 1;
+                    int iForward = Math.Min(propertyIndex, count);
+                    int iBackward = iForward - 1;
                     while (iForward < count || iBackward >= 0)
                     {
                         if (iForward < count)
                         {
-                            if (TryIsPropertyRefEqual(ref _propertyRefsSorted[iForward], propertyName, key, ref info))
+                            if (TryIsPropertyRefEqual(_propertyRefsSorted[iForward], propertyName, key, ref info))
                             {
                                 return info;
                             }
@@ -233,7 +225,7 @@ namespace System.Text.Json
 
                         if (iBackward >= 0)
                         {
-                            if (TryIsPropertyRefEqual(ref _propertyRefsSorted[iBackward], propertyName, key, ref info))
+                            if (TryIsPropertyRefEqual(_propertyRefsSorted[iBackward], propertyName, key, ref info))
                             {
                                 return info;
                             }
@@ -246,60 +238,77 @@ namespace System.Text.Json
             // Try the main list which has all of the properties in a consistent order.
             // We could get here even when hasPropertyCache==true if there is a race condition with different json
             // property ordering and _propertyRefsSorted is re-assigned while in the loop above.
-            for (int i = 0; i < _propertyRefs.Count; i++)
-            {
-                PropertyRef propertyRef = _propertyRefs[i];
-                if (TryIsPropertyRefEqual(ref propertyRef, propertyName, key, ref info))
-                {
-                    break;
-                }
-            }
 
-            if (!hasPropertyCache)
-            {
-                if (propertyIndex == 0 && frame.PropertyRefCache == null)
-                {
-                    // Create the temporary list on first property access to prevent a partially filled List.
-                    frame.PropertyRefCache = new HashSet<PropertyRef>();
-                }
+            string stringPropertyName = Encoding.UTF8.GetString(propertyName
+#if netstandard
+                        .ToArray()
+#endif
+            );
 
-                if (info != null)
+            if (PropertyCache.TryGetValue(stringPropertyName, out info))
+            {
+                // For performance, only add to cache up to a threshold and then just use the dictionary.
+                int count;
+                if (_propertyRefsSorted != null)
                 {
-                    Debug.Assert(frame.PropertyRefCache != null);
-                    frame.PropertyRefCache.Add(new PropertyRef(key, info));
+                    count = _propertyRefsSorted.Length;
+                }
+                else
+                {
+                    count = 0;
+                }
+                
+                // Do a quick check for the stable (after warm-up) case.
+                if (count < PropertyNameCountCacheThreshold)
+                {
+                    if (frame.PropertyRefCache != null)
+                    {
+                        count += frame.PropertyRefCache.Count;
+                    }
+
+                    // Check again to fill up to the limit.
+                    if (count < PropertyNameCountCacheThreshold)
+                    {
+                        if (frame.PropertyRefCache == null)
+                        {
+                            frame.PropertyRefCache = new List<PropertyRef>();
+                        }
+
+                        ulong key = info.PropertyNameKey;
+                        PropertyRef propertyRef = new PropertyRef(key, info);
+                        frame.PropertyRefCache.Add(propertyRef);
+                    }
                 }
             }
 
             return info;
         }
 
-        internal JsonPropertyInfo GetPolicyProperty()
+        private Dictionary<string, JsonPropertyInfo> CreatePropertyCache(int capacity)
         {
-            Debug.Assert(_propertyRefs.Count == 1);
-            return _propertyRefs[0].Info;
-        }
+            StringComparer comparer;
 
-        internal JsonPropertyInfo GetProperty(int index)
-        {
-            Debug.Assert(index < _propertyRefs.Count);
-            return _propertyRefs[index].Info;
-        }
-
-        internal int PropertyCount
-        {
-            get
+            if (Options.PropertyNameCaseInsensitive)
             {
-                return _propertyRefs.Count;
+                comparer = StringComparer.OrdinalIgnoreCase;
             }
+            else
+            {
+                comparer = StringComparer.Ordinal;
+            }
+
+            return new Dictionary<string, JsonPropertyInfo>(capacity, comparer);
         }
 
-        private static bool TryIsPropertyRefEqual(ref PropertyRef propertyRef, ReadOnlySpan<byte> propertyName, ulong key, ref JsonPropertyInfo info)
+        public JsonPropertyInfo PolicyProperty { get; private set; }
+
+        private static bool TryIsPropertyRefEqual(in PropertyRef propertyRef, ReadOnlySpan<byte> propertyName, ulong key, ref JsonPropertyInfo info)
         {
             if (key == propertyRef.Key)
             {
                 if (propertyName.Length <= PropertyNameKeyLength ||
                     // We compare the whole name, although we could skip the first 6 bytes (but it's likely not any faster)
-                    propertyName.SequenceEqual(propertyRef.Info.NameUsedToCompare))
+                    propertyName.SequenceEqual(propertyRef.Info.Name))
                 {
                     info = propertyRef.Info;
                     return true;
@@ -323,7 +332,7 @@ namespace System.Text.Json
             return false;
         }
 
-        private static ulong GetKey(ReadOnlySpan<byte> propertyName)
+        public static ulong GetKey(ReadOnlySpan<byte> propertyName)
         {
             ulong key;
             int length = propertyName.Length;