… src/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializer.Write.HandleDictionary.cs
… src/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializer.Write.HandleEnumerable.cs
… src/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializer.Write.HandleObject.cs
… src/System.Text.Json/src/System/Text/Json/Serialization/ReadStackFrame.cs
… src/System.Text.Json/src/System/Text/Json/Serialization/WriteStackFrame.cs