Changes the way properties are looked up to improve performance, security and handle case-insensitive in all cultures. The perf improvement help most when case insensitive is enabled, but also helps with the default case-sensitive case.

Fixes https://github.com/dotnet/corefx/issues/38510
Fixes https://github.com/dotnet/corefx/issues/36939

~4% improvement for case-sensitive.
~27% improvement for case-insensitive.

Before (simple parent\child POCO model):
```
|                         Method |      Mean |     Error |    StdDev |   Median |      Min |       Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------------- |----------:|----------:|----------:|---------:|---------:|----------:|------------:|------------:|------------:|--------------------:|
|                       JSON.NET |  84.66 us |  3.588 us |  3.840 us | 83.20 us | 80.16 us |  90.79 us |      7.4983 |      1.0225 |           - |            31.31 KB |
|               System.Text.Json |  65.57 us |  1.796 us |  1.922 us | 64.77 us | 63.08 us |  69.36 us |      8.3879 |      0.2621 |           - |            34.82 KB |
| System.Text.Json.CaseSensitive |  90.45 us |  4.321 us |  4.977 us | 88.75 us | 83.91 us | 101.02 us |     15.1569 |      0.7050 |           - |            62.44 KB |
```

After
```
|------------------------------- |---------:|---------:|---------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|                       JSON.NET | 82.14 us | 2.578 us | 2.969 us | 80.76 us | 79.03 us | 88.00 us |      7.3671 |      0.9609 |           - |            31.31 KB |
|               System.Text.Json | 63.12 us | 1.879 us | 2.089 us | 62.59 us | 60.55 us | 66.36 us |      8.3089 |      0.4888 |           - |            34.82 KB |
| System.Text.Json.CaseSensitive | 66.43 us | 2.298 us | 2.646 us | 65.22 us | 64.13 us | 73.25 us |      9.8420 |      0.5180 |           - |            41.13 KB |
```

Note: there will soon be additional tests for Turkish characters and POCOs with >64 properties.