@steveharter 2019-06-26
Committing this; some misc threading questions will be discussed offline.